backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-multipart
python-dotenv
pillow
sqlalchemy
alembic     
scikit-learn
joblib